[#11] pipe와 exporter 계층 통합
